+++ src/backend/ci/core/worker/worker-agent/build.gradle
@@ -31,8 +31,6 @@ mainClassName = "com.tencent.devops.agent.ApplicationKt"
 dependencies {
     compile project(":core:worker:worker-common")
     compile project(":core:worker:worker-api-sdk")
-    compile project(":core:worker:worker-plugin-scm")
-    compile project(":core:worker:worker-plugin-archive")
 
     compile project (":core:plugin:codecc-plugin:worker-plugin-codecc")
