+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/plugin/worker/task/scm/CodeSvnPullTask.kt
@@ -24,34 +24,11 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.plugin.worker.task
+package com.tencent.devops.plugin.worker.task.scm
 
-import com.tencent.devops.common.api.exception.ParamBlankException
-import com.tencent.devops.common.archive.element.SingleArchiveElement
-import com.tencent.devops.process.pojo.BuildTask
-import com.tencent.devops.process.pojo.BuildVariables
-import com.tencent.devops.worker.common.task.ITask
+import com.tencent.devops.common.api.enums.ScmType
+import com.tencent.devops.common.pipeline.pojo.element.agent.CodeSvnElement
 import com.tencent.devops.worker.common.task.TaskClassType
-import com.tencent.devops.worker.common.utils.ArchiveUtils.archiveCustomFiles
-import com.tencent.devops.worker.common.utils.ArchiveUtils.archivePipelineFiles
-import java.io.File
 
-@TaskClassType(classTypes = [SingleArchiveElement.classType])
-class SingleFileArchiveTask : ITask() {
-
-    override fun execute(buildTask: BuildTask, buildVariables: BuildVariables, workspace: File) {
-        val taskParams = buildTask.params ?: mapOf()
-        val filePath = taskParams["filePath"] ?: throw ParamBlankException("param [filePath] is empty")
-        val isCustomize = taskParams["customize"] ?: throw ParamBlankException("param [isCustomize] is empty")
-
-        val count = if (isCustomize.toBoolean()) {
-            val destPath = taskParams["destPath"] ?: throw ParamBlankException("param [destPath] is empty")
-            archiveCustomFiles(filePath, destPath, workspace, buildVariables)
-        } else {
-            archivePipelineFiles(filePath, workspace, buildVariables)
-        }
-        if (count == 0) {
-            throw RuntimeException("没有匹配到任何待归档文件，请检查工作空间下面的文件")
-        }
-    }
-}
+@TaskClassType(classTypes = [CodeSvnElement.classType])
+class CodeSvnPullTask : CodePullTask(ScmType.CODE_SVN)
