+++ src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/plugin/worker/task/scm/git/DefaultGitPullCodeSetting.kt
